Move func into a utils file
